Add files via upload
Ask to the user to enter the age of a kid then determine his/her category (Demander à l'utilisateur d'entrer l'âge d'un enfant puis déterminer sa catégorie): It was my first interrogation on the way to create an algorithm in Algobox. We wrote our algorithms on  sheets of paper.